Disable drupal core dev branch testing for now.
It would be nice to have this to verify that the site works in newest
versions of Drupal. But it's not possible to get a clean test in this
branch atm because of a bug with drupal-scaffold and composer, that I'm
not able to resolve right now. Look here for more information:

https://github.com/drupal-composer/drupal-scaffold/issues/79

Try to redo the tests on next update by reverting this commit.