Revert "Use our config/sync directory"
After updating Drupal core to 8.6.1 the method of using the
config_install profile on CI to verify was failing. The config sync
validation is suddenly failing with the reason that the standard profile
is missing as a module. So this reverts it temporarily, until we can
figure out what's going on.