Update installers plugin and set appropiate patch level for core.
In the new version of installers extender composer plugin it's possible
to set patch level for individual projects. We use this to set the
appropiate level for core, since drupal/core only packages a subdir of
the upstream drupal/drupal project, which patches are generated from.

See: https://github.com/drupal-composer/drupal-project/issues/412